+++ deploy/auto_cron.sh
@@ -1,18 +1,18 @@
 #!/bin/sh
 set -euo pipefail
 
-source /home/webapps/miniconda3/bin/activate auto_qc
+source /home/autoqc/miniconda3/bin/activate auto_qc
 
-cd /export/home/webapps/auto_qc 
+cd /u01/apps/autoqc/auto_qc/
 
 
-echo doing the wren novaseq
-python manage.py update_database --raw_data_dir /mnt/wren_archive/quality_temp/novaseq/ --config /export/home/webapps/auto_qc/config/config_gen01.yaml
+echo -e INFO"\t"$(date +%Y-%m-%d_%H-%M-%S)"\t"doing the wren novaseq
+python manage.py update_database --raw_data_dir /mnt/wren/wren_archive/quality_temp/novaseq/ --config /u01/apps/autoqc/auto_qc/config/config_webserver.yaml
 
-echo doing the wren miseq
-python manage.py update_database --raw_data_dir /mnt/wren_archive/quality_temp/miseq/ --config /export/home/webapps/auto_qc/config/config_gen01.yaml
+echo -e INFO"\t"$(date +%Y-%m-%d_%H-%M-%S)"\t"doing the wren miseq
+python manage.py update_database --raw_data_dir /mnt/wren/wren_archive/quality_temp/miseq/ --config /u01/apps/autoqc/auto_qc/config/config_webserver.yaml
 
-echo doing the wren nextseq
-python manage.py update_database --raw_data_dir /mnt/wren_archive/quality_temp/nextseq/ --config /export/home/webapps/auto_qc/config/config_gen01.yaml
+echo -e INFO"\t"$(date +%Y-%m-%d_%H-%M-%S)"\t"doing the wren nextseq
+python manage.py update_database --raw_data_dir /mnt/wren/wren_archive/quality_temp/nextseq/ --config /u01/apps/autoqc/auto_qc/config/config_webserver.yaml
 
-source /home/webapps/miniconda3/bin/deactivate
+source /home/autoqc/miniconda3/bin/deactivate