+++ mysite/settings.py
@@ -12,12 +12,12 @@ https://docs.djangoproject.com/en/2.2/ref/settings/
 
 import os
 
-deploy_location = 'gen01'
+deploy_location = 'webserver'
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
-CONFIG_PATH = 'config/config.yaml'
+CONFIG_PATH = 'config/config_webserver.yaml'
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/2.2/howto/deployment/checklist/
 
@@ -27,7 +27,7 @@ SECRET_KEY = 'r5()y@mri3mn@*24#$n#h2jdehru85g@6vo$oe975nd@(h7ql*'
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
-ALLOWED_HOSTS = ['127.0.0.1', '10.59.210.245']
+ALLOWED_HOSTS = ['127.0.0.1', '10.59.210.245', '10.69.115.27']
 
 
 # Application definition
@@ -78,9 +78,11 @@ WSGI_APPLICATION = 'mysite.wsgi.application'
 
 # Database
 
-if deploy_location == 'gen01':
+if deploy_location == 'webserver':
 
-	DB_PASSWORD_FILE = '/export/home/webapps/password.txt'
+	URL_PREFIX = "autoqc/"
+
+	DB_PASSWORD_FILE = '/u01/apps/autoqc/password.txt'
 	with open(DB_PASSWORD_FILE) as f:
 		db_password = f.readline().strip()
 
@@ -97,6 +99,8 @@ if deploy_location == 'gen01':
 	}
 
 else:
+
+	URL_PREFIX = ""
 
 	DATABASES = {
 	'default': {
@@ -136,7 +140,13 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/2.2/howto/static-files/
 
-STATIC_URL = '/static/'
+if deploy_location == "webserver":
+	STATIC_URL = '/static/autoqc/'
+	STATIC_ROOT = '/var/www/static/autoqc'
+
+else:
+	STATIC_URL = '/static/'
+
 STATICFILES_DIRS = [
 	os.path.join(BASE_DIR, "static"),
 ]
@@ -144,7 +154,7 @@ STATICFILES_DIRS = [
 CRISPY_TEMPLATE_PACK = 'bootstrap4'
 
 LOGIN_REDIRECT_URL = 'home_auto_qc'
-LOGIN_URL = '/login/'
+LOGIN_URL = f'/{URL_PREFIX}login/'
 
 
 MESSAGE_SLACK = False
@@ -185,3 +195,12 @@ with open(HPO_FILEPATH) as file:
 			alt_id = line.split('alt_id:')[1].strip()
 			
 			HPO_TERMS_DICT[alt_id] = name
+
+# SampleSheet generator download location
+if deploy_location == 'gen01':
+
+	SSGEN_DOWNLOAD = '/mnt/wren/wren_temp/SampleSheets/'
+
+else:
+
+	SSGEN_DOWNLOAD = '/home/awmgs/SampleSheets/'